requirements.txt
# omni/som
torch>=2.2.1
torchvision>=0.17.1
ultralytics>=8.1.28
easyocr>=1.7.1
numpy>=1.26.4
pillow>=10.2.0
setuptools>=75.8.1
opencv-python-headless>=4.11.0.86
matplotlib>=3.8.3
huggingface-hub>=0.21.4
supervision>=0.25.1
typing-extensions>=4.9.0
pydantic>=2.6.3
fastapi[standard]

google-adk
playwright
rich
termcolor
litellm
opik